requirements.txt
# Django core
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg[binary]>=3.1,<4.0

# Cache & message broker
redis>=5.0,<6.0

# Async / WebSocket support
channels>=4.0,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0

# Background tasks
celery>=5.4,<6.0
django-celery-beat>=2.6,<3.0
django-celery-results>=2.5,<3.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Exchange connectors
python-binance>=1.0.19,<2.0
alpaca-py>=0.30,<1.0
websockets>=13.0,<14.0

# Data & analysis
numpy>=1.26,<3.0
pandas>=2.2,<3.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-asyncio>=0.23,<1.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0
responses>=0.25,<1.0
aioresponses>=0.7,<1.0

# Code quality
ruff>=0.4,<1.0
